requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==8.0.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
dateparser==1.1.0
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
frozenlist==1.2.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.2.0
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.15.0
SQLAlchemy==1.3.18
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.7
websockets==9.1
Werkzeug==1.0.1
yarl==1.7.2
